Export operations specific to mesaides to clean json file
